+++ src/components/file_objects/folder.rs
@@ -158,7 +158,7 @@ impl FileObject for Folder {
 
     fn generate_outline(
         &self,
-        depth: u32,
+        depth: u64,
         export_string: &mut String,
         objects: &super::FileObjectStore,
     ) {
@@ -178,7 +178,7 @@ impl FileObject for Folder {
 
     fn generate_export(
         &self,
-        depth: u32,
+        depth: u64,
         export_string: &mut String,
         objects: &super::FileObjectStore,
         export_options: &ExportOptions,
